+++ src/lda.cpp
@@ -1,3 +1,4 @@
+#define ARMA_NO_DEBUG
 #include "lib.h"
 #include "dev.h"
 #include "lda.h"
@@ -8,11 +9,12 @@ using namespace Rcpp;
 
 // [[Rcpp::export]]
 List cpp_lda(arma::sp_mat &mt, int k, int max_iter, double alpha, double beta,
-                  arma::sp_mat &seeds, int seed, bool verbose) {
+                  arma::sp_mat &seeds, arma::sp_mat &words, int random, bool verbose) {
     LDA lda;
-    lda.set_data(mt);
-    lda.seed = seed;
     lda.K = k;
+    lda.set_data(mt);
+    lda.set_fitted(words);
+    lda.random = random;
     if (max_iter > 0)
         lda.niters = max_iter;
     if (alpha > 0)
@@ -26,7 +28,7 @@ List cpp_lda(arma::sp_mat &mt, int k, int max_iter, double alpha, double beta,
         arma::umat s;
         if (seeded) {
             if (arma::size(seeds) != arma::size(lda.nw))
-                std::invalid_argument("Invalid seed matrix");
+                throw std::invalid_argument("Invalid seed matrix");
             s = arma::conv_to<arma::umat>::from(arma::mat(seeds));
             lda.nw = lda.nw + s; // set pseudo count
             //lda.nwsum = lda.nwsum + arma::sum(s, 0);
@@ -39,9 +41,11 @@ List cpp_lda(arma::sp_mat &mt, int k, int max_iter, double alpha, double beta,
     lda.compute_phi();
 
     return List::create(Rcpp::Named("k") = lda.K,
-                        Rcpp::Named("iter") = lda.liter,
+                        Rcpp::Named("max_iter") = lda.niters,
+                        Rcpp::Named("last_iter") = lda.liter,
                         Rcpp::Named("alpha") = lda.alpha,
                         Rcpp::Named("beta") = lda.beta,
                         Rcpp::Named("phi") = wrap(lda.phi),
-                        Rcpp::Named("theta") = wrap(lda.theta));
+                        Rcpp::Named("theta") = wrap(lda.theta),
+                        Rcpp::Named("words") = wrap(arma::sp_umat(lda.nw)));
 }